LayoutTests/fast/events/ios/meaningful-click-when-focusing-body-expected.txt

This test exercises the 'meaningful click' heuristic in an editable body element, and requires WebKitTestRunner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Focused body element
PASS didDispatchNonMeaningfulClickCallback is false
PASS Tapped inside body element
PASS didDispatchNonMeaningfulClickCallback is false
PASS successfullyParsed is true

TEST COMPLETE
